requirements.txt
flask==3.0.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
requests==2.31.0
gunicorn==22.0.0
gTTS==2.5.1
